SQUASH ME: respond to comments
