+++ environs/imagemetadata/simplestreams.go
@@ -157,11 +157,6 @@ const (
 	ReleasedStream = "released"
 )
 
-// ImageRelease maps a base to an image version.
-func ImageRelease(base corebase.Base) (string, error) {
-	return base.Channel.Track, nil
-}
-
 // idStream returns the string to use in making a product id
 // for the given product stream.
 func idStream(stream string) string {
