+++ internal/provider/azure/userdata.go
@@ -16,10 +16,8 @@ import (
 type AzureRenderer struct{}
 
 func (AzureRenderer) Render(cfg cloudinit.CloudConfig, os ostype.OSType) ([]byte, error) {
-	switch os {
-	case ostype.Ubuntu:
-		return renderers.RenderYAML(cfg, utils.Gzip, renderers.ToBase64)
-	default:
+	if os != ostype.Ubuntu {
 		return nil, errors.Errorf("Cannot encode userdata for OS: %s", os)
 	}
+	return renderers.RenderYAML(cfg, utils.Gzip, renderers.ToBase64)
 }
