+++ internal/provider/maas/userdata.go
@@ -16,10 +16,8 @@ import (
 type MAASRenderer struct{}
 
 func (MAASRenderer) Render(cfg cloudinit.CloudConfig, os ostype.OSType) ([]byte, error) {
-	switch os {
-	case ostype.Ubuntu:
-		return renderers.RenderYAML(cfg, utils.Gzip, renderers.ToBase64)
-	default:
-		return nil, errors.Errorf("Cannot encode userdata for OS: %s", os.String())
+	if os != ostype.Ubuntu {
+		return nil, errors.Errorf("Cannot encode userdata for OS: %s", os)
 	}
+	return renderers.RenderYAML(cfg, utils.Gzip, renderers.ToBase64)
 }
