+++ internal/provider/oci/images.go
@@ -257,19 +257,13 @@ func getImageType(img ociCore.Image) ImageType {
 // NewInstanceImage returns a populated InstanceImage from the ociCore.Image
 // struct returned by oci's API, the image's architecture or an error.
 func NewInstanceImage(img ociCore.Image, compartmentID *string) (InstanceImage, string, error) {
-	var (
-		err       error
-		arch      string
-		base      corebase.Base
-		isMinimal bool
-		imgType   InstanceImage
-	)
-	switch osName := *img.OperatingSystem; osName {
-	case ubuntuOS:
-		base, arch, isMinimal = parseUbuntuImage(img)
-	default:
+	if osName := *img.OperatingSystem; osName != ubuntuOS {
 		return InstanceImage{}, "", errors.NotSupportedf("os %s", osName)
 	}
+	var (
+		imgType InstanceImage
+	)
+	base, arch, isMinimal := parseUbuntuImage(img)
 
 	imgType.ImageType = getImageType(img)
 	imgType.Id = *img.Id
